Move Core Data to production
