+++ EssentialDeveloper/EssentialDeveloper/Feed Cache/CoreData/CoreDataFeedStore.swift
@@ -1,0 +1,60 @@
+import CoreData
+
+public final class CoreDataFeedStore: FeedStore {
+    private let container: NSPersistentContainer
+    private let context: NSManagedObjectContext
+    
+    public init(storeURL: URL, bundle: Bundle = .main) throws {
+        container = try NSPersistentContainer.load(modelName: "FeedStore",url: storeURL, in: bundle)
+        context = container.newBackgroundContext()
+    }
+    
+    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        let context = self.context
+        context.perform {
+            do {
+                try ManagedCache.findAndDestroy(in: context)
+                try context.save()
+
+                completion(nil)
+            }catch {
+                completion(error)
+            }
+        }
+    }
+    
+    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        let context = self.context
+        context.perform {
+            do {
+                let managedCache = try ManagedCache.createNewUniqueInstance(in: context)
+                managedCache.timestamp = timestamp
+                managedCache.feed = ManagedFeedImage.images(from: feed, in: context)
+                
+                try context.save()
+                completion(nil)
+            } catch {
+                completion(error)
+            }
+        }
+    }
+    
+    public func retrieve(completion: @escaping RetrievalCompletion) {
+        let context = self.context
+        context.perform {
+            do {
+                if let cache = try ManagedCache.find(in: context) {
+                    completion(.found(
+                        feed: cache.localFeed,
+                        timestamp: cache.timestamp
+                    ))
+                } else {
+                    completion(.empty)
+                }
+            } catch {
+                completion(.failure(error))
+            }
+        }
+    }
+}
+
